sv_ttk_colorizer/requirements.txt
darkdetect
Pillow
requests
pywinstyles
sv_ttk
appdirs